requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-groq
langchain-huggingface
sentence-transformers
chromadb
python-dotenv
requests
pydantic
streamlit
requests